Nueva Policy y separacion de codigo en nuevo usuario

+++ app/Http/Controllers/API/TipoVehiculoController.php
@@ -5,11 +5,16 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\Models\TipoVehiculo;
 use Illuminate\Http\Request;
-
 use App\Http\Resources\TipoVehiculoResource;
 
 class TipoVehiculoController extends Controller
 {
+
+    public function __construct()
+    {
+        // $this->authorizeResource(TipoVehiculo::class, 'tipoVehiculos');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -43,6 +48,7 @@ class TipoVehiculoController extends Controller
      */
     public function show(TipoVehiculo $tipoVehiculo)
     {
+        $this->authorize('view',$tipoVehiculo);
         return new TipoVehiculoResource($tipoVehiculo);
     }
 
@@ -55,6 +61,7 @@ class TipoVehiculoController extends Controller
      */
     public function update(Request $request, TipoVehiculo $tipoVehiculo)
     {
+        $this->authorize('update',$tipoVehiculo);
         $tipoVehiculoData = json_decode($request->getContent(), true);
         $tipoVehiculo->update($tipoVehiculoData['data']['attributes']);
 
@@ -69,6 +76,7 @@ class TipoVehiculoController extends Controller
      */
     public function destroy(TipoVehiculo $tipoVehiculo)
     {
+        $this->authorize('delete',$tipoVehiculo);
         $tipoVehiculo->delete();
     }
 }
